move getLabelProbabilities to SongCollection
